(#37 prefs) add: distribute upper even of the given n of points for each dimension
Represent N using a triplet of dimensions - hard cube roots found recursively. For example, 12 is (3, 2, 2) and 18 is (3, 3, 2), 99 is (5, 5, 4) where the product of dimensions is the closest larger number whose cube root is an integer. Distributed vertices are expected to be shuffled and excessive ones are ignored. This approach always ensures an even distribution of 'right N' vertices.